move provider files into cdap package
