+++ cdap/resource_application.go
@@ -12,24 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package cdap
 
 import (
 	"encoding/json"
 	"net/http"
-	"os"
 	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
 
-// https://docs.cdap.io/cdap/current/en/reference-manual/http-restful-api/artifact.html
-func resourceArtifact() *schema.Resource {
+// https://docs.cdap.io/cdap/current/en/reference-manual/http-restful-api/lifecycle.html.
+func resourceApplication() *schema.Resource {
 	return &schema.Resource{
-		Create: resourceArtifactCreate,
-		Read:   resourceArtifactRead,
-		Delete: resourceArtifactDelete,
-		Exists: resourceArtifactExists,
+		Create: resourceApplicationCreate,
+		Read:   resourceApplicationRead,
+		Delete: resourceApplicationDelete,
+		Exists: resourceApplicationExists,
 
 		Schema: map[string]*schema.Schema{
 			"name": {
@@ -45,20 +44,7 @@ func resourceArtifact() *schema.Resource {
 					return defaultNamespace, nil
 				},
 			},
-			"version": {
-				Type:     schema.TypeString,
-				Optional: true,
-				ForceNew: true,
-			},
-			"extends": {
-				Type: schema.TypeList,
-				Elem: &schema.Schema{
-					Type: schema.TypeString,
-				},
-				Optional: true,
-				ForceNew: true,
-			},
-			"jar_binary_path": {
+			"config": {
 				Type:     schema.TypeString,
 				Required: true,
 				ForceNew: true,
@@ -67,32 +53,17 @@ func resourceArtifact() *schema.Resource {
 	}
 }
 
-func resourceArtifactCreate(d *schema.ResourceData, m interface{}) error {
+func resourceApplicationCreate(d *schema.ResourceData, m interface{}) error {
 	config := m.(*Config)
 	name := d.Get("name").(string)
-	addr := urlJoin(config.host, "/v3/namespaces", d.Get("namespace").(string), "/artifacts", name)
 
-	jar, err := os.Open(d.Get("jar_binary_path").(string))
+	addr := urlJoin(config.host, "/v3/namespaces", d.Get("namespace").(string), "/apps", name)
+
+	body := strings.NewReader(d.Get("config").(string))
+
+	req, err := http.NewRequest(http.MethodPut, addr, body)
 	if err != nil {
 		return err
-	}
-	defer jar.Close()
-
-	req, err := http.NewRequest(http.MethodPost, addr, jar)
-	if err != nil {
-		return err
-	}
-
-	req.Header = map[string][]string{}
-	if v, ok := d.GetOk("version"); ok {
-		req.Header.Add("Artifact-Version", v.(string))
-	}
-	if v, ok := d.GetOk("extends"); ok {
-		var es []string
-		for _, e := range v.([]interface{}) {
-			es = append(es, e.(string))
-		}
-		req.Header.Add("Artifact-Extends", strings.Join(es, "/"))
 	}
 
 	if _, err := httpCall(config.client, req); err != nil {
@@ -103,14 +74,14 @@ func resourceArtifactCreate(d *schema.ResourceData, m interface{}) error {
 	return nil
 }
 
-func resourceArtifactRead(d *schema.ResourceData, m interface{}) error {
+func resourceApplicationRead(d *schema.ResourceData, m interface{}) error {
 	return nil
 }
 
-func resourceArtifactDelete(d *schema.ResourceData, m interface{}) error {
+func resourceApplicationDelete(d *schema.ResourceData, m interface{}) error {
 	config := m.(*Config)
 	name := d.Get("name").(string)
-	addr := urlJoin(config.host, "/v3/namespaces", d.Get("namespace").(string), "/artifacts", name, "/versions", d.Get("version").(string))
+	addr := urlJoin(config.host, "/v3/namespaces", d.Get("namespace").(string), "/apps", name)
 
 	req, err := http.NewRequest(http.MethodDelete, addr, nil)
 	if err != nil {
@@ -120,11 +91,10 @@ func resourceArtifactDelete(d *schema.ResourceData, m interface{}) error {
 	return err
 }
 
-func resourceArtifactExists(d *schema.ResourceData, m interface{}) (bool, error) {
+func resourceApplicationExists(d *schema.ResourceData, m interface{}) (bool, error) {
 	config := m.(*Config)
 	name := d.Get("name").(string)
-	addr := urlJoin(config.host, "/v3/namespaces", d.Get("namespace").(string), "/artifacts")
-
+	addr := urlJoin(config.host, "/v3/namespaces", d.Get("namespace").(string), "/apps")
 	req, err := http.NewRequest(http.MethodGet, addr, nil)
 	if err != nil {
 		return false, err
@@ -135,16 +105,16 @@ func resourceArtifactExists(d *schema.ResourceData, m interface{}) (bool, error)
 		return false, err
 	}
 
-	type artifact struct {
+	type app struct {
 		Name string `json:"name"`
 	}
 
-	var artifacts []artifact
-	if err := json.Unmarshal(b, &artifacts); err != nil {
+	var apps []app
+	if err := json.Unmarshal(b, &apps); err != nil {
 		return false, err
 	}
 
-	for _, a := range artifacts {
+	for _, a := range apps {
 		if a.Name == name {
 			return true, nil
 		}
