+++ main.go
@@ -16,6 +16,8 @@
 package main
 
 import (
+	"terraform-provider-cdap/cdap"
+
 	"github.com/hashicorp/terraform-plugin-sdk/plugin"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 )
@@ -23,7 +25,7 @@ import (
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() terraform.ResourceProvider {
-			return Provider()
+			return cdap.Provider()
 		},
 	})
 }
