--- resource_cdap_namespace.go
@@ -1,107 +1,0 @@
-// Copyright 2020 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"encoding/json"
-	"net/http"
-
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-)
-
-// https://docs.cdap.io/cdap/current/en/reference-manual/http-restful-api/namespace.html
-func resourceNamespace() *schema.Resource {
-	return &schema.Resource{
-		Create: resourceNamespaceCreate,
-		Read:   resourceNamespaceRead,
-		Delete: resourceNamespaceDelete,
-		Exists: resourceNamespaceExists,
-
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
-			},
-		},
-	}
-}
-
-func resourceNamespaceCreate(d *schema.ResourceData, m interface{}) error {
-	config := m.(*Config)
-	name := d.Get("name").(string)
-	addr := urlJoin(config.host, "/v3/namespaces", name)
-
-	req, err := http.NewRequest(http.MethodPut, addr, nil)
-	if err != nil {
-		return err
-	}
-
-	if _, err := httpCall(config.client, req); err != nil {
-		return err
-	}
-
-	d.SetId(name)
-	return nil
-}
-
-func resourceNamespaceRead(d *schema.ResourceData, m interface{}) error {
-	return nil
-}
-
-func resourceNamespaceDelete(d *schema.ResourceData, m interface{}) error {
-	config := m.(*Config)
-	name := d.Get("name").(string)
-	addr := urlJoin(config.host, "/v3/namespaces", name)
-
-	req, err := http.NewRequest(http.MethodDelete, addr, nil)
-	if err != nil {
-		return err
-	}
-	_, err = httpCall(config.client, req)
-	return err
-}
-
-func resourceNamespaceExists(d *schema.ResourceData, m interface{}) (bool, error) {
-	config := m.(*Config)
-	name := d.Get("name").(string)
-	addr := urlJoin(config.host, "/v3/namespaces")
-
-	req, err := http.NewRequest(http.MethodGet, addr, nil)
-	if err != nil {
-		return false, err
-	}
-
-	b, err := httpCall(config.client, req)
-	if err != nil {
-		return false, err
-	}
-
-	type namespace struct {
-		Name string `json:"name"`
-	}
-
-	var namespaces []namespace
-	if err := json.Unmarshal(b, &namespaces); err != nil {
-		return false, err
-	}
-
-	for _, n := range namespaces {
-		if n.Name == name {
-			return true, nil
-		}
-	}
-	return false, nil
-}
